requirements.txt
python-dotenv

fastapi
uvicorn[standard]
jinja2

python-telegram-bot
